python-sidecar/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
schedule==1.2.1
requests==2.32.5
scikit-learn==1.6.1
numpy==2.2.1
pandas==2.2.3